backend/pyproject.toml
[project]
name = "Kartik"
version = "0.1.0"
description = "RAG"
authors = [
    { name = "Kartik" }
]
requires-python = ">=3.8"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "python-dotenv",
    "pydantic",
    "google-generativeai",
    "pinecone",
    "cohere",
    "numpy",
    "PyMuPDF",
]

[tool.uv]
dev-dependencies = []